pyproject.toml
[project]
name = "427-assignment1-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "rti-connext>=7.6.0",
]
